LayoutTests/fast/css/will-change-parsing-expected.txt

This tests checks that will-change parses correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyle("will-change:;") is "auto"
PASS testComputedStyle("will-change: contents, all;") is "auto"
PASS testComputedStyle("will-change: contents;") is "contents"
PASS test("will-change: inherit;") is "inherit"
PASS test("will-change: initial;") is "initial"
FAIL test("will-change: all;") should be  (of type string). Was null (of type object).
FAIL test("will-change: none;") should be  (of type string). Was null (of type object).
FAIL test("will-change: default;") should be  (of type string). Was null (of type object).
FAIL test("will-change: will-change;") should be  (of type string). Was null (of type object).
PASS test("will-change: auto;") is "auto"
FAIL test("will-change: auto, contents;") should be  (of type string). Was null (of type object).
PASS test("will-change: scroll-position;") is "scroll-position"
PASS test("will-change: contents;") is "contents"
PASS test("will-change: scroll-position, scroll-position;") is "scroll-position, scroll-position"
PASS test("will-change: contents, scroll-position, transform;") is "contents, scroll-position, transform"
PASS test("will-change: contents, transform, scroll-position, transform;") is "contents, transform, scroll-position, transform"
FAIL test("will-change: , background;") should be  (of type string). Was null (of type object).
PASS test("will-change: background, transform;") is "background, transform"
FAIL test("will-change: background transform;") should be  (of type string). Was null (of type object).
FAIL test("will-change: background, all;") should be  (of type string). Was null (of type object).
FAIL test("will-change: background, will-change;") should be  (of type string). Was null (of type object).
PASS test("will-change: unknown-property;") is "unknown-property"
FAIL test("will-change: transform, initial;") should be  (of type string). Was null (of type object).
FAIL test("will-change: default;") should be  (of type string). Was null (of type object).
PASS successfullyParsed is true

TEST COMPLETE
